+++ elena/elena_bot.py
@@ -3,102 +3,245 @@ import os
 
 import numpy as np
 
-from elena.exchange import Exchange, OrderStatus
+from elena.exchange_manager import ExchangeManager, OrderStatus
 from elena.logging import llog
 from elena.utils import get_time
 
 
 class Elena:
-    def __init__(self, robot_filename: str, exchange: Exchange):
+    def __init__(self, robot_filename: str, exchange: ExchangeManager):
         self._robot_filename = robot_filename
         self._exchange = exchange
+        self._state = self._read_state()
+
+    def _reset_state(self):
+        self._state['buy'] = 0.0
+        self._state['buy_order_id'] = ''
+        self._state['buy_order'] = None
+        self._state['sell'] = 0.0
+        self._state['sell_order_id'] = ''
+        self._state['sell_order'] = None
+        self._state['status'] = ''
+        self._state['iteration_benefit'] = 0.0
+        self._state['iteration_margin'] = 0.0
+        self._state['iteration_current_bid'] = 0
+        self._state['iteration_current_ask'] = 0
+        self._state['iteration_current_value'] = 0
+        self._state['iteration_current_value_USD'] = 0
+        self._state['iteration_current_margin'] = 0
+        self._state['iteration_current_total_benefit'] = 0
+        self._state['left_on_asset'] = 0.0
+        self._state['sleep_until'] = 0
+        self._state['sell_status'] = None
 
     def iterate(self):
-        state = self._read_state()
-        if state['sleep_until'] < get_time():
-
-            if not state['buy_order_id'] and state['active']:
-                buy, sell = self._estimate_buy_sel(state)
+        if self._state['sleep_until'] < get_time():
+            if not self._state['buy_order_id'] and self._state['active']:
+                self._parameters_migration()
+                buy, sell = self._estimate_buy_sel()
                 if buy > 0:
                     llog("create a new buy order")
-                    new_buy_order_id = self._exchange.create_buy_order(state['max_order'], state['symbol'], buy)
-                    # TODO: create a reset_state
-                    state['sleep_until'] = 0
-                    state['buy_order_id'] = new_buy_order_id
-                    state['sell_order_id'] = ''
-                    state['buy'] = buy
-                    state['sell'] = sell
-                    self._save_state(state)
+                    force_buy_price = self._state['algo'] > 7
+                    new_buy_order = self._exchange.create_buy_order(self._state['max_order'], self._state['symbol'], buy, force_buy_price)
+                    self._reset_state()
+                    self._state['buy_order_id'] = new_buy_order['orderId']
+                    self._state['buy_order'] = new_buy_order
+                    self._state['buy'] = buy
+                    self._state['sell'] = sell
+                    self._state['status'] = 'buying'
+                    self._update_orders_status_values_and_profits(force_update_buy_order=True)
+                    self._save_state()
                 else:
-                    state['sleep_until'] = self._sleep_until(get_time(), 5)
-                    self._save_state(state)
+                    self._state['sleep_until'] = self._sleep_until(get_time(), 5)
+                    self._state['status'] = "waiting - don't buy"
+                    self._save_state()
                     llog("don't buy")
-                return
+                # return --- maybe the order is executed immediately
 
-            if state['buy_order_id'] and not state['sell_order_id']:
-                new_sell_order_id = self._exchange.create_sell_order(state['symbol'], state['buy_order_id'],
-                                                                     state['sell'])
-                if new_sell_order_id:
-                    llog("create a new sell order")
-                    state['sell_order_id'] = new_sell_order_id
-                    self._save_state(state)
-                else:
-                    status, order_update_time = self._exchange.check_order_status(state['symbol'],
-                                                                                  state['buy_order_id'])
-                    if not status == OrderStatus.CANCELED.value:
-                        state['sleep_until'] = self._sleep_until(get_time(), 5)
-                        self._save_state(state)
-                        llog("waiting purchase")
+            if self._state['buy_order_id'] and not self._state['sell_order_id']:
+                self._update_orders_status_values_and_profits(force_update_buy_order=True)
+                buy_order = self._state['buy_order']
+                status = buy_order['status']
+                order_time = int(buy_order['time'])
+                order_age_limit = order_time + (float(self._state['buy_auto_cancel_timeout']) * 60 * 1000)
+                now = get_time()
+                if status == OrderStatus.FILLED.value:
+                    self._create_sell_order(self._state['sell'])
+                elif status == OrderStatus.CANCELED.value:
+                    llog("buy cancellation detected")
+                    self._add_cycles()
+                    self._save_history()
+                    if self._state['active'] == 1:
+                        self._reset_state()
+                        self._state['status'] = 'buy cancellation'
+                        self._save_state()
                     else:
-                        llog("buy cancellation")
-                        self._save_history_state_and_profit(state)
-                        if state['active'] == 1:
-                            state['sleep_until'] = 0
-                            state['buy_order_id'] = ''
-                            state['sell_order_id'] = ''
-                            state['buy'] = 0
-                            state['sell'] = 0
-                            self._save_state(state)
-                        else:
-                            self._delete_state()
-                return
+                        self._delete_state()
+                elif status == OrderStatus.NEW.value and now > order_age_limit:
+                    llog("auto buy cancellation")
+                    cancellation = self._exchange.cancel_order(self._state['symbol'], self._state['buy_order_id'])
+                    llog(cancellation)
+                    self._state['status'] = 'auto buy cancellation'
+                    self._add_buy_auto_cancel_count()
+                    self._save_state()
+                else:
+                    # self._state['sleep_until'] = self._sleep_until(get_time(), 2)
+                    self._save_state()
+                    llog("waiting purchase")
+                # return --- maybe the order is executed immediately
 
-            if state['buy_order_id'] and state['sell_order_id']:
-                status, order_update_time = self._exchange.check_order_status(state['symbol'], state['sell_order_id'])
+            if self._state['buy_order_id'] and self._state['sell_order_id']:
+                self._update_orders_status_values_and_profits()
+                sell_order = self._state['sell_order']
+                status = sell_order['status']
+                sell_order_update_time = int(sell_order['updateTime'])
+                order_age_limit = sell_order_update_time + (float(self._state['sell_auto_cancel_timeout']) * 60 * 1000)
+                now = get_time()
                 if status == OrderStatus.FILLED.value:
                     llog("save history")
-                    self._save_history_state_and_profit(state)
-                    if state['active'] == 1:
+                    self._add_sales()
+                    self._save_history()
+                    if self._state['active'] == 1:
                         llog("set sleep")
-                        state['sleep_until'] = self._sleep_until(order_update_time, state['data_samples'] * 1.5)
-                        state['buy_order_id'] = ''
-                        state['sell_order_id'] = ''
-                        state['buy'] = 0
-                        state['sell'] = 0
-                        self._save_state(state)
+                        self._calculate_reinvest_or_losses()
+                        self._reset_state()
+                        self._state['sleep_until'] = self._sleep_until(sell_order_update_time, self._state['data_samples'] * self._state['sleep_until_factor'])
+                        self._state['status'] = 'waiting'
+                        self._save_state()
                     else:
+                        self._save_state()
                         self._delete_state()
                 elif status == OrderStatus.CANCELED.value:
                     llog("sell cancellation, save history")
-                    self._save_history_state_and_profit(state)
+                    self._state['status'] = 'manual sell cancellation'
+                    self._add_cycles()
+                    self._save_history()
+                    self._save_state()
                     self._delete_state()
+                elif status == OrderStatus.NEW.value and now > order_age_limit and float(self._state['sell_auto_cancel_timeout'])>0:
+                    self._state['sell_status'] = "sell timeout"
+                    buy_order = self._state['buy_order']
+                    order_buy_price = float(buy_order['price'])  # get the real price when bought
+                    order_sell_price = float(sell_order['price'])  # get the sell price as is in the exchange
+
+                    bid = self._state['iteration_current_bid']
+                    minimum_profit = 1 + ((self._exchange._minimum_profit - 1) / 2)
+                    minimum_price = order_buy_price * minimum_profit
+
+                    if bid >= minimum_price:  # earning
+                        if self._state['sell_auto_cancel_im_feeling_lucky_data_samples'] > 0:
+                            buy, sell = self._estimate_buy_sel(data_samples=self._state['sell_auto_cancel_im_feeling_lucky_data_samples'])
+                            if sell > order_sell_price:
+                                buy_order_id = self._state['buy_order_id']
+                                text = f'Cancel and sell at higher price, ' \
+                                       f'order_id:{buy_order_id}, ' \
+                                       f'order_buy_price:{order_buy_price}, ' \
+                                       f'order_sell_price:{order_sell_price}, new_sell_price:{sell}'
+                                llog(text)
+                                self._state['sell_status'] = text
+                                self._cancel_sell_order_and_create_a_new_one(sell)
+                                return
+                            elif sell > 0:
+                                llog("Sell time out but do nothing. set sleep to 0.")
+                                self._state['sell_status'] = "Sell time out but do nothing. Tendence is high."
+                                self._save_state()
+                                return
+                            else:
+                                pass  # Cancel order and sell at bid (we get some benefit)
+                        llog("Cancel order and sell at bid (we get some benefit)")
+                        self._state['sell_status'] = "Cancel order and sell at bid (we get some benefit)"
+                        self._cancel_sell_order_and_create_a_new_one(bid, force_sell_price=True)
+                    else:  # loosing
+                        percentage_loss = self._state['iteration_current_margin'] * -1
+                        absolute_loss = (order_buy_price - bid) * float(sell_order['origQty'])
+                        text = f'loss%:{percentage_loss}, order_buy_price:{order_buy_price}, bid:{bid}, loss$:{absolute_loss}'
+                        # llog(text)
+                        self._state['sell_status'] = text
+                        self._save_state()
+
+                        stop_loss_percentage_absolute = (
+                                self._state['stop_loss_percentage_absolute'] > 0 and
+                                percentage_loss <= self._state['stop_loss_percentage_absolute'])
+
+                        stop_loss_percentage_relative_to_accumulated_benefits = (
+                                self._state['stop_loss_percentage_relative_to_accumulated_benefits'] > 0 and
+                                absolute_loss <= (self._state['accumulated_benefit'] * (self._state['stop_loss_percentage_relative_to_accumulated_benefits'] / 100)))
+
+                        if stop_loss_percentage_absolute or stop_loss_percentage_relative_to_accumulated_benefits:
+                            text = f'Cancel order and sell at bid (losing:{percentage_loss}%,{absolute_loss}$)'
+                            llog(text)
+                            self._state['sell_status'] = text
+                            self._cancel_sell_order_and_create_a_new_one(bid, force_sell_price=True)
                 else:
-                    state['sleep_until'] = self._sleep_until(get_time(), 15)
-                    self._save_state(state)
+                    self._state['sleep_until'] = self._sleep_until(get_time(), 5)
+                    self._save_state()
                     llog("waiting sell")
-                return
+
+                # return --- maybe the order is executed immediately
         else:
             llog("sleeping")
 
-    def _read_state(self):
+    @staticmethod
+    def _verify_key_set_default(state, key, default_value):
+        if state.get(key) is None:
+            state[key] = default_value
+
+    @staticmethod
+    def _del_key(state, key):
+        if not state.get(key) is None:
+            del state[key]
+
+    def _parameters_migration(self):
+        # should be done only before a new cycle starts
+        if self._state['algo'] == 4 or self._state['algo'] == 6 or self._state['algo'] == 8:
+            self._state['algo'] = 11
+        if self._state['algo'] == 5 or self._state['algo'] == 7 or self._state['algo'] == 9:
+            self._state['algo'] = 12
+        if self._state['algo'] == 10:
+            self._state['algo'] = 13
+
+    def _read_state(self):  # TODO: if it uses self._robot_filename why is not using self._state?
         fp = open(self._robot_filename, 'r')
         state = json.load(fp)
         fp.close()
+
+        # default values
+        self._verify_key_set_default(state, 'buy_order_id', 0)
+        self._verify_key_set_default(state, 'sell_order_id', 0)
+        self._verify_key_set_default(state, 'sleep_until', 0)
+        self._verify_key_set_default(state, 'sleep_until_factor', 1.5)
+
+        self._verify_key_set_default(state, 'buy_auto_cancel_timeout', 5)
+        self._verify_key_set_default(state, 'buy_auto_cancel_count', 0)
+        self._verify_key_set_default(state, 'iteration_current_bid', 0)
+        self._verify_key_set_default(state, 'iteration_current_ask', 0)
+        self._verify_key_set_default(state, 'iteration_current_value', 0)
+        self._verify_key_set_default(state, 'iteration_current_value_USD', 0)
+        self._verify_key_set_default(state, 'iteration_current_margin', 0)
+        self._verify_key_set_default(state, 'iteration_current_total_benefit', 0)
+
+        self._verify_key_set_default(state, 'sell_auto_cancel_timeout', 5)
+        self._verify_key_set_default(state, 'sell_auto_cancel_count', 0)
+        self._verify_key_set_default(state, 'sell_auto_cancel_im_feeling_lucky_data_samples', 10)
+        self._verify_key_set_default(state, 'sell_auto_cancel_lucky_count', 0)
+        self._verify_key_set_default(state, 'stop_loss_percentage_absolute', 0)
+        self._verify_key_set_default(state, 'stop_loss_percentage_relative_to_accumulated_benefits', 0)
+        self._verify_key_set_default(state, 'reinvest', 0)
+        self._verify_key_set_default(state, 'accumulated_benefit', 0)
+        self._verify_key_set_default(state, 'accumulated_benefit_USD', 0)
+        self._verify_key_set_default(state, 'accumulated_margin', 0)
+        self._verify_key_set_default(state, 'sales', 0)
+        self._verify_key_set_default(state, 'cycles', 0)
+
+        # migrations
+
         return state
 
-    def _save_state(self, state, filename=None):
+    def _save_state(self, state=None, filename=None):
         if not filename:
             filename = self._robot_filename
+        if not state:
+            state = self._state
         fp = open(filename, 'w')
         json.dump(state, fp)
         fp.close()
@@ -106,45 +249,128 @@ class Elena:
     def _delete_state(self):
         os.rename(self._robot_filename, self._robot_filename + '.inactive')
 
-    def _save_history_state_and_profit(self, in_state):
-        state = dict(in_state)
-        filename = f"history/{str(get_time())}_{str(state['buy_order_id'])}.json"
+    def _add_cycles(self):
+        self._state['cycles'] = self._state['cycles'] + 1
+
+    def _add_sales(self):
+        self._add_cycles()
+        self._state['sales'] = self._state['sales'] + 1
+
+    def _add_buy_auto_cancel_count(self):
+        self._state['buy_auto_cancel_count'] = self._state['buy_auto_cancel_count'] + 1
+
+    def _add_sell_auto_cancel_count(self):
+        self._state['sell_auto_cancel_count'] = self._state['sell_auto_cancel_count'] + 1
+
+    def _add_sell_auto_cancel_lucky_count(self):
+        self._state['sell_auto_cancel_lucky_count'] = self._state['sell_auto_cancel_lucky_count'] + 1
+
+    def _update_orders_status_values_and_profits(self, force_update_buy_order=False):
         buy_order = ''
         sell_order = ''
         iteration_benefit = 0
         iteration_margin = 0
         left_on_asset = 0
 
-        if state['buy_order_id']:
-            buy_order = self._exchange.get_order(state['buy_order_id'], state['symbol'] )
-            if state['sell_order_id']:
-                sell_order = self._exchange.get_order(state['sell_order_id'], state['symbol'])
+        if self._state['buy_order_id']:
+            check_order = True
+            if self._state['buy_order']:
+                buy_order = self._state['buy_order']
+                status = buy_order['status']
+                check_order = not (status == OrderStatus.FILLED.value) or force_update_buy_order
+            if check_order:
+                buy_order = self._exchange.get_order(self._state['buy_order_id'], self._state['symbol'])
+                self._state['buy_order'] = buy_order
+
+            if self._state['sell_order_id']:
+                sell_order = self._exchange.get_order(self._state['sell_order_id'], self._state['symbol'])
+                self._state['sell_order'] = sell_order
                 if sell_order['status'] == OrderStatus.FILLED.value:
                     iteration_benefit = float(sell_order['cummulativeQuoteQty']) - float(buy_order['cummulativeQuoteQty'])
                     iteration_margin = (iteration_benefit / float(buy_order['cummulativeQuoteQty'])) * 100
                     left_on_asset = float(buy_order['executedQty']) - float(sell_order['executedQty'])
 
-        state['active'] = -1
-        state['buy_order'] = buy_order
-        state['sell_order'] = sell_order
-        state['iteration_benefit'] = iteration_benefit
-        state['iteration_margin'] = iteration_margin
-        state['left_on_asset'] = left_on_asset
+                if sell_order['status'] == OrderStatus.NEW.value:
+                    bid, ask = self._exchange.get_cached_order_book_first_bids_asks(self._state['symbol'])
+                    self._state['iteration_current_bid'] = bid
+                    self._state['iteration_current_ask'] = ask
+
+                    buy_order = self._state['buy_order']
+                    order_buy_price = float(buy_order['price'])  # get the real price when bought
+                    self._state['iteration_current_margin'] = (bid / order_buy_price) * 100 - 100
+
+                    self._state['iteration_current_value'] = bid * float(buy_order['origQty'])
+                    self._state['iteration_current_value_USD'] = self._exchange.convert_to_usd(self._state['symbol'], self._state['iteration_current_value'])
+                    self._state['iteration_current_total_benefit'] = self._state['accumulated_benefit'] + self._state['iteration_current_value'] - self._state['max_order']
+
+        self._state['iteration_benefit'] = iteration_benefit
+        self._state['iteration_margin'] = iteration_margin
+        self._state['left_on_asset'] = left_on_asset
+
+        if iteration_benefit != 0:
+            self._state['accumulated_benefit'] = self._state['accumulated_benefit'] + iteration_benefit
+            self._state['accumulated_benefit_USD'] = self._exchange.convert_to_usd(self._state['symbol'], self._state['accumulated_benefit_USD'])
+            self._state['accumulated_margin'] = self._state['accumulated_margin'] + iteration_margin
+            if iteration_benefit < 0:
+                llog("iteration margin <0!", self._state)
+
+    def _save_history(self):
+        self._update_orders_status_values_and_profits()
+
+        history_state = dict(self._state)
+        filename = f"history/{str(get_time())}_{str(history_state['buy_order_id'])}.json"
+
+        history_state['active'] = -1
 
         fp = open(filename, 'w')
-        json.dump(state, fp)
+        json.dump(history_state, fp)
         fp.close()
 
-    def _estimate_buy_sel(self, state):
-        candles_df = self._exchange.get_candles(p_symbol=state['symbol'], p_limit=state['data_samples'])
-        return self._buy_sell(candles_df, state['algo'], state['margin'], state['tendence_tolerance'])
+
+
+    def _calculate_reinvest_or_losses(self):
+        # re-invest
+        if self._state['reinvest'] > 0 and self._state['iteration_benefit'] > 0:
+            self._state['max_order'] = self._state['max_order'] + (self._state['iteration_benefit'] * self._state['reinvest'] / 100)
+
+        # losses
+        if self._state['iteration_benefit'] < 0:
+            self._state['max_order'] = self._state['max_order'] + self._state['iteration_benefit']
+
+    def _create_sell_order(self, sell, force_sell_price=False):
+        buy_order = self._state['buy_order']
+        sell_quantity = float(buy_order['executedQty'])
+        new_sell_order = self._exchange.create_sell_order(self._state['symbol'], sell_quantity, sell, force_sell_price)
+        if new_sell_order:
+            llog("created a new sell order", new_sell_order)
+            self._state['sell_order_id'] = new_sell_order['orderId']
+            self._state['sell_order'] = new_sell_order
+            self._state['status'] = 'selling'
+            self._save_state()
+
+    def _cancel_sell_order_and_create_a_new_one(self, sell, force_sell_price=False):
+        cancellation = self._exchange.cancel_order(self._state['symbol'], self._state['sell_order_id'])
+        llog(cancellation)
+        if sell < self._state['sell']:
+            self._add_sell_auto_cancel_count()
+        else:
+            self._add_sell_auto_cancel_lucky_count()
+        self._state['sell_order_id'] = 0
+        self._save_state()  # TODO: review if it's necessary... if the order was canceled but the new one can't be executed in the next iteration this would be understood as a human cancelation.
+        self._create_sell_order(sell, force_sell_price)
+
+    def _estimate_buy_sel(self, data_samples=None):
+        if data_samples is None:
+            data_samples = self._state['data_samples']
+        candles_df = self._exchange.get_candles(p_symbol=self._state['symbol'], p_limit=data_samples)
+        return self._buy_sell(candles_df, self._state['algo'], self._state['margin'], self._state['tendence_tolerance'])
 
     @staticmethod
     def _sleep_until(sell_execution_time, minutes):
         return sell_execution_time + minutes * 60 * 1000  # 45' after the sale
 
     @staticmethod
-    def _ensure_sell_is_higher_than_buy_by(next_sell, next_buy, minimum_profit=1.0005):
+    def _ensure_sell_is_higher_than_buy_by(next_sell, next_buy, minimum_profit):
         # ensure sell is higher than buy at least by 5 per thousand to pay fees
         if next_sell / next_buy > minimum_profit:
             buy = next_buy
@@ -155,19 +381,51 @@ class Elena:
         return buy, sell
 
     @staticmethod
-    def _buy_sell_based_on_linear_regression(candles_df_buy_sell, sell_field, margin=0):
+    def _sell_based_on_linear_regression(candles_df_buy_sell, sell_field, margin=0):
+        # TODO: reusing margin but is the number of steps to extrapolate
+        regression_close = np.polyfit(candles_df_buy_sell.index, candles_df_buy_sell[sell_field], 1)
+        return regression_close[0] * (candles_df_buy_sell.index[-1] + 1 + margin) + regression_close[1]
+
+    def _buy_sell_based_on_linear_regression(self, candles_df_buy_sell, sell_field, margin=0):
         regression_low = np.polyfit(candles_df_buy_sell.index, candles_df_buy_sell["Low"], 1)
         next_low = regression_low[0] * (candles_df_buy_sell.index[-1] + 1) + regression_low[1]
 
-        # TODO: reusing margin but is the number of steps to extrapolate
-        regression_close = np.polyfit(candles_df_buy_sell.index, candles_df_buy_sell[sell_field], 1)
-        next_close = regression_close[0] * (candles_df_buy_sell.index[-1] + 1 + margin) + regression_close[1]
+        next_close = self._sell_based_on_linear_regression(candles_df_buy_sell, sell_field, margin=margin)
 
-        buy, sell = Elena._ensure_sell_is_higher_than_buy_by(next_close, next_low)
+        buy, sell = Elena._ensure_sell_is_higher_than_buy_by(next_close, next_low, self._exchange._minimum_profit)
         return buy, sell
 
-    @staticmethod
-    def _buy_sell(candles_df_buy_sell, algo, margin, tendence_tolerance):
+    def _buy_on_bid_sell_based_on_linear_regression(self, candles_df_buy_sell, sell_field, margin=0):
+        next_low, ask = self._exchange.get_cached_order_book_first_bids_asks(self._state['symbol'])
+
+        next_close = self._sell_based_on_linear_regression(candles_df_buy_sell, sell_field, margin=margin)
+
+        buy, sell = Elena._ensure_sell_is_higher_than_buy_by(next_close, next_low, self._exchange._minimum_profit)
+        return buy, sell
+
+    def _buy_on_bid_sell_based_on_fixed_margin(self, margin):
+        buy, ask = self._exchange.get_cached_order_book_first_bids_asks(self._state['symbol'])
+
+        sell = buy * (1 + (margin / 100))
+
+        return buy, sell
+
+    def _buy_on_ask_sell_based_on_linear_regression(self, candles_df_buy_sell, sell_field, margin=0):
+        bid, next_low = self._exchange.get_cached_order_book_first_bids_asks(self._state['symbol'])
+
+        next_close = self._sell_based_on_linear_regression(candles_df_buy_sell, sell_field, margin=margin)
+
+        buy, sell = Elena._ensure_sell_is_higher_than_buy_by(next_close, next_low, self._exchange._minimum_profit)
+        return buy, sell
+
+    def _buy_on_ask_sell_based_on_fixed_margin(self, margin):
+        bid, buy = self._exchange.get_cached_order_book_first_bids_asks(self._state['symbol'])
+
+        sell = buy * (1 + (margin / 100))
+
+        return buy, sell
+
+    def _buy_sell(self, candles_df_buy_sell, algo, margin, tendence_tolerance):
         sell = 0
         buy = 0
 
@@ -190,14 +448,26 @@ class Elena:
                 buy = buy.astype(float)
                 sell = buy * (1 + (margin / 100))
             if algo == 4:
-                buy, sell = Elena._buy_sell_based_on_linear_regression(candles_df_buy_sell, "Close", margin=0)
+                buy, sell = self._buy_sell_based_on_linear_regression(candles_df_buy_sell, "Close", margin=0)
             if algo == 5:
-                buy, sell = Elena._buy_sell_based_on_linear_regression(candles_df_buy_sell, "High", margin=0)
+                buy, sell = self._buy_sell_based_on_linear_regression(candles_df_buy_sell, "High", margin=0)
             if algo == 6:
-                buy, sell = Elena._buy_sell_based_on_linear_regression(candles_df_buy_sell, "Close", margin=margin)
+                buy, sell = self._buy_sell_based_on_linear_regression(candles_df_buy_sell, "Close", margin=margin)
             if algo == 7:
-                buy, sell = Elena._buy_sell_based_on_linear_regression(candles_df_buy_sell, "High", margin=margin)
+                buy, sell = self._buy_sell_based_on_linear_regression(candles_df_buy_sell, "High", margin=margin)
+            if algo == 8:
+                buy, sell = self._buy_on_bid_sell_based_on_linear_regression(candles_df_buy_sell, "Close", margin=margin)
+            if algo == 9:
+                buy, sell = self._buy_on_bid_sell_based_on_linear_regression(candles_df_buy_sell, "High", margin=margin)
+            if algo == 10:
+                buy, sell = self._buy_on_ask_sell_based_on_fixed_margin(margin)
+            if algo == 11:
+                buy, sell = self._buy_on_ask_sell_based_on_linear_regression(candles_df_buy_sell, "Close", margin=margin)
+            if algo == 12:
+                buy, sell = self._buy_on_ask_sell_based_on_linear_regression(candles_df_buy_sell, "High", margin=margin)
+            if algo == 13:
+                buy, sell = self._buy_on_bid_sell_based_on_fixed_margin(margin)
         if algo == 3:
-            buy, sell = Elena._buy_sell_based_on_linear_regression(candles_df_buy_sell, "Close", margin=0)
+            buy, sell = self._buy_sell_based_on_linear_regression(candles_df_buy_sell, "Close", margin=0)
 
         return buy, sell